feat: use shadcn badge, refactor components
