+++ app/blog/[slug]/page.tsx
@@ -1,8 +1,8 @@
 import { notFound } from "next/navigation";
-import { CustomMDX } from "app/components/mdx";
+import { CustomMDX } from "@/components/composite/mdx";
 import { formatDate, getBlogPosts } from "app/blog/utils";
 import { BASE_URL } from "app/sitemap";
-import { BadgeLink } from "app/components/badge-link";
+import { BadgeLink } from "@/components/composite/badge-link";
 import { FaArrowLeft } from "react-icons/fa6";
 
 export async function generateStaticParams() {
