+++ app/blog/page.tsx
@@ -1,5 +1,5 @@
-import { BadgeLink } from "app/components/badge-link";
-import { BlogPosts } from "app/components/posts";
+import { BadgeLink } from "@/components/composite/badge-link";
+import { BlogPosts } from "@/components/composite/posts";
 import { FaArrowLeft } from "react-icons/fa6";
 
 export const metadata = {
