+++ app/layout.tsx
@@ -3,10 +3,10 @@ import { Rubik } from "next/font/google";
 import type { Metadata } from "next";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import Footer from "./components/footer";
+import Footer from "../components/composite/footer";
 import { BASE_URL } from "./sitemap";
 import { DESCRIPTION, TITLE } from "./constants";
-import Header from "./components/header";
+import Header from "../components/composite/header";
 import { cn } from "@/lib/utils";
 
 const rubik = Rubik({ subsets: ["latin"] });
