+++ components/composite/profile-picture/profile-picture.tsx
